requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-genai
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.40.0
